updated the folder structure
