+++ src/styles/fonts.ts
@@ -1,0 +1,63 @@
+import {
+  Inter,
+  Inika,
+  Alata,
+  Judson,
+  Kaisei_Opti,
+  DM_Sans,
+  Amaranth,
+} from "next/font/google";
+
+export const f_inter_800 = Inter({
+  weight: "800",
+  subsets: ["latin"],
+  display: "swap",
+});
+
+export const f_inter_700 = Inter({
+  weight: "700",
+  subsets: ["latin"],
+  display: "swap",
+});
+
+export const f_inter_400 = Inter({
+  weight: "400",
+  subsets: ["latin"],
+  display: "swap",
+});
+
+export const f_inika_700 = Inika({
+  weight: "700",
+  subsets: ["latin"],
+  display: "swap",
+});
+
+export const f_alata_400 = Alata({
+  weight: "400",
+  subsets: ["latin"],
+  display: "swap",
+});
+
+export const f_judson_700 = Judson({
+  weight: "700",
+  subsets: ["latin"],
+  display: "swap",
+});
+
+export const f_kaiseiOpti_700 = Kaisei_Opti({
+  weight: "700",
+  subsets: ["latin"],
+  display: "swap",
+});
+
+export const f_dmsans_700 = DM_Sans({
+  weight: "700",
+  subsets: ["latin"],
+  display: "swap",
+});
+
+export const f_amaranth_400 = Amaranth({
+  weight: "400",
+  subsets: ["latin"],
+  display: "swap",
+});
