+++ src/ui/ComponentExporter.ts
@@ -1,7 +1,1 @@
 export { default as HomeLayout } from "./Home/Home";
-export { default as Navbar } from "./Navbar/Navbar";
-export { default as MastHead } from "./Masthead/Masthead";
-export { default as Info } from "./Info/Info";
-export { default as Sidebar } from "./Sidebar/Sidebar";
-export { default as Slider } from "./Slider/Slider";
-export { default as Footer } from "./Footer/Footer";
